requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environ==1.0
executing==1.2.0
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.8.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.0.2
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3
traitlets==5.9.0
tzdata==2023.4
wcwidth==0.2.6
